fix: 차트 버그수정
